refactor : 유저 데코레이터 변경
